fairplay: revised door open animation to allow reveals underneath.
Nearly identical to observer. Instead of glyph_L moving left
and glyph_R moving right, which leaves undrawn gaps between door
edge and reveal, I put glyph_R on left and vice versa so that the
gaps inherent in those glyphs are part of the DOOR instead of the
space they leave behind.